requirements.txt
python-telegram-bot==13.4.1
SQLAlchemy==1.4.7
python-dotenv==0.17.0